switch to selectors over takeLatest
