+++ src/store/selectors/chartRange.ts
@@ -1,0 +1,17 @@
+import { format } from 'date-fns';
+import { createSelector } from 'starfx/store';
+
+import { schema } from '~/src/store/schema.ts';
+
+export const dateRangeWithStrings = createSelector(
+  schema.chartRange.select,
+  (graphRange) => {
+    const { start, end } = graphRange;
+    return {
+      start,
+      startString: format(start, 'yyyy-MM-dd'),
+      end,
+      endString: format(end, 'yyyy-MM-dd')
+    };
+  }
+);
