+++ src/store/thunks/index.ts
@@ -2,6 +2,6 @@ export { thunks } from './foundation.ts';
 export * from './accounts.ts';
 export * from './settings.ts';
 export * from './transactions.ts';
-import { tasks as chartTasks } from './chartData.ts';
+export * from './chartRange.ts';
 
-export const tasks = [...chartTasks];
+export const tasks = [];
